Ds new 16072026 feature

+++ Backend/backend.Core/DTOs/WarehouseDetailsDto.cs
@@ -8,7 +8,7 @@ public class WarehouseDetailsDto
     public string? WarehouseName { get; set; }
 
     public string? ApplicationIdNo { get; set; }
-    public long RegId { get; set; }
+    public int RegId { get; set; }
 
     public string CatCode { get; set; } = string.Empty;
     public string? WarehouseAddress1 { get; set; }
